move queries.
