+++ src/main/kotlin/com/openlattice/graph/Graph.kt
@@ -1190,14 +1190,6 @@ private val EDGE_IDS_SQL = "${EDGE_ENTITY_KEY_ID.name} = ANY(?) AND ${EDGE_ENTIT
 private val SRC_IDS_AND_ENTITY_SETS_SQL = "${SRC_ENTITY_KEY_ID.name} = ANY(?) AND ${SRC_ENTITY_SET_ID.name} = ANY(?)"
 private val DST_IDS_AND_ENTITY_SETS_SQL = "${DST_ENTITY_KEY_ID.name} = ANY(?) AND ${DST_ENTITY_SET_ID.name} = ANY(?)"
 
-private val REPARTITION_SELECTOR = "partitions[ 1 + ((array_length(partitions,1) + (('x'||right(${SRC_ENTITY_KEY_ID.name}::text,8))::bit(32)::int % array_length(partitions,1))) % array_length(partitions,1))]"
-private val REPARTITION_COLUMNS = E.columns.joinToString(",") { if (it == PARTITION) REPARTITION_SELECTOR else it.name }
-private val REPARTITION_SQL = """
-INSERT INTO ${E.name} SELECT $REPARTITION_COLUMNS
-    FROM ${E.name} INNER JOIN (select id as entity_set_id, partitions from entity_sets) as es 
-    ON ${SRC_ENTITY_SET_ID.name} = es.entity_set_id
-    WHERE entity_set_id = ? AND partition!=$REPARTITION_SELECTOR
-""".trimIndent()
 //private val REPARTITION_SQL = "INSERT INTO ${E.name} SELECT $REPARTITION_COLUMNS FROM ${E.name} INNER JOIN (select id as entity_set_id, partitions from entity_sets) as es using (entity_set_id) WHERE entity_set_id = ? AND "
 
 /**
